+++ docutils/HISTORY.rst
@@ -17,6 +17,13 @@
 Release 0.22.3b1.dev (unpublished)
 ==================================
 
+* docutils/parsers/rst/directives/body.py,
+  docutils/parsers/rst/directives/parts.py
+
+  - Ignore auxiliary elements like `sphinx.addnodes.only`
+    when testing for invalid parents of topics or sidebars
+    (cf. https://github.com/sphinx-doc/sphinx/issues/14002).
+
 * docutils/parsers/rst/states.py
 
   - Ignore combining characters when extracting a grid table block